pyproject.toml
[tool.poetry]
name = "creator"
version = "0.1.0"
description = ""
authors = ["vijayantajain <vijayantajain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
pydantic = "^2.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
